+++ src/main/java/kitchenpos/menu/application/response/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.application.response;
 
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuGroupResponse {
