+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,10 +1,10 @@
-package kitchenpos.menu.application;
+package kitchenpos.menugroup.application;
 
 import java.util.stream.Collectors;
 import kitchenpos.menu.application.request.MenuGroupCreateRequest;
 import kitchenpos.menu.application.response.MenuGroupResponse;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.repository.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
