+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -9,11 +9,11 @@ import kitchenpos.config.ServiceTest;
 import kitchenpos.menu.application.request.MenuCreateRequest;
 import kitchenpos.dto.ProductQuantityDto;
 import kitchenpos.menu.application.response.MenuResponse;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.menu.repository.MenuGroupRepository;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import kitchenpos.product.repository.ProductRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
