+++ src/test/java/kitchenpos/menu/repository/MenuGroupRepositoryTest.java
@@ -3,8 +3,9 @@ package kitchenpos.menu.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.config.RepositoryTest;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
